additional fixes to arc - still some droped pixels in the middle of the arc

+++ sgl/src/sgl_renderer.hpp
@@ -28,7 +28,7 @@ struct SglRenderer
     void push_vertex(const SglVertex & vertex);
     void draw_circle(const SglVertex & center, int radius);
     void draw_ellipse(const SglVertex & center, int a, int b, SglMatrix mat);
-    void draw_arc(const SglVertex & center, int radius, float from, float to, SglMatrix mat);
+    void draw_arc(const SglVertex & center, float radius, float from, float to, SglMatrix mat);
 
     private:
         friend struct SglCore;
